fix(mailchimp): remove circular dependency on errors
