+++ packages/backend-modules/mail/errors.js
@@ -1,8 +1,5 @@
-const NEWSLETTER_MEMBER_ERROR = 'NEWSLETTER_MEMBER_ERROR'
 const INTERESTID_NOT_FOUND_ERROR = 'INTERESTID_NOT_FOUND_ERROR'
-const EMAIL_REQUIRED_ERROR = 'EMAIL_REQUIRED_ERROR'
 const SUBSCRIPTION_CONFIG_MISSING_ERROR = 'SUBSCRIPTION_CONFIG_MISSING_ERROR'
-const SUBSCRIPTION_HANDLER_MISSING_ERROR = 'SUBSCRIPTION_HANDLER_MISSING_ERROR'
 const SEND_ERROR = 'SEND_ERROR'
 
 class MailError extends Error {
@@ -20,30 +17,11 @@ class SubscriptionConfigurationMissingMailError extends MailError {
   }
 }
 
-class SubscriptionHandlerMissingMailError extends MailError {
-  constructor(meta) {
-    super(SUBSCRIPTION_HANDLER_MISSING_ERROR, meta)
-  }
-}
-
-class NewsletterMemberMailError extends MailError {
-  constructor(meta) {
-    super(NEWSLETTER_MEMBER_ERROR, meta)
-  }
-}
-
 class InterestIdNotFoundMailError extends MailError {
   constructor(meta) {
     super(INTERESTID_NOT_FOUND_ERROR, meta)
   }
 }
-
-class EmailRequiredMailError extends MailError {
-  constructor(meta) {
-    super(EMAIL_REQUIRED_ERROR, meta)
-  }
-}
-
 class SendMailError extends MailError {
   constructor(meta) {
     super(SEND_ERROR, meta)
@@ -51,10 +29,7 @@ class SendMailError extends MailError {
 }
 
 module.exports = {
-  NewsletterMemberMailError,
   InterestIdNotFoundMailError,
-  EmailRequiredMailError,
   SubscriptionConfigurationMissingMailError,
-  SubscriptionHandlerMissingMailError,
   SendMailError,
 }
